requirements.txt
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0
requests>=2.31.0
python-3.11.0
flask==3.0.2
gunicorn==21.2.0